# swagger/swagger.yaml
...
info:
  contact: {}
  description: 请输入以 `Bearer ` 开头的 Token，例：Bearer xxxxxxxx。未列出接口请参考前端调用方法。Token在个人中心-API密钥菜单下申请。
  title: k8m API
  version: "1.0"
paths:
...
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/configmap:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的ConfigMap
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/endpoints:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的端点
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/env:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的环境变量
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/envFromPod:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的来自其他Pod的环境变量
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/ingress:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的Ingress
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/node:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的节点
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/pv:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的PV
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/pvc:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的PVC
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/secret:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的Secret
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/ns/{ns}/name/{name}/links/services:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: API组
        in: path
        name: group
        required: true
        type: string
      - description: API版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Pod关联的服务
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"